requirements.txt
aiohttp-requests
numpy
cmake
pydantic~=1.8.2
opencv-python
requests
aioredis
pyaml
Pillow
pandas
openpyxl
asyncpgsa~=0.27.1
sqlalchemy~=1.3.20

redis
psycopg2-binary